pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "FastAPI Auth0 example application"
authors = ["Michael Oliver <contact@michaeloliver.dev>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.2"
uvicorn = "^0.14.0"
httpx = "^0.18.1"
python-dotenv = "^0.17.1"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
black = { version = "^21.5b2", allow-prereleases = true }
ipython = "^7.24.1"
mypy = "^0.901"
flake8 = "^3.9.2"
isort = "^5.8.0"
pydocstyle = "^6.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pydocstyle]
convention = "google"
add_ignore = ["D100"]

[tool.mypy]
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "jose"
ignore_missing_imports = true
